postgres_to_es/src/requirements.txt
psycopg2-binary==2.8.6
python-dotenv==0.15.0
requests==2.25.0